Change date output to ISO 8601 format in sovisuhal module
The datetime objects, specifically "date_to", were previously converted to strings using the default 'date()' method. This commit modifies that by using the 'isoformat()' method to ensure the output is in ISO 8601 format. This change has been made across different files in the 'sovisuhal' module to maintain consistency. The new format improves compatibility with other services and software as the ISO 8601 is internationally recognized.