Add affiliations handling in check view and template
Added several changes to handle affiliations in 'view_cbv.py'. Implemented a new method 'get_affiliations_case' to fetch details of affiliations and a remove function to delete an existing affiliation from the list. This is reflected in the template files where a new section for Affiliations is added. This enhancement provides users with an interface to manage affiliations more effectively.