Add new template for structure index and update existing template for searcher index
The change includes the creation of a new template file for the structure indexing feature, as well as modifications to the existing searcher indexing feature to suit new requirements. Refactoring was also done to better accommodate for multiple indexing categories in structure and searcher index. The headers and menus have also been updated to adequately reflect the changes.