Add LDAP ID to navigation links in templates
Extended navigation link URLs in 'structures_index.html' and 'searchers_index.html' to include the 'ldapid' parameter. This was necessary because certain functionalities need an LDAP ID to be passed along with the URL. The change ensures that the specified LDAP ID will now be include in the addressed URL whenever available.