Improve data filtering by affiliation and adjust comment formatting
Refactored methods in 'view_cbv.py' to filter data based on affiliation, increasing efficiency and simplifying the templates. Introduced 'sv_affiliation' as a parameter in 'get_elastic_data()' and 'get_searcher_list()' to directly fetch the data required. The condition checking for the affiliation inside the entity list iteration in 'searchers_index.html' is removed as filtering is now handled within the data fetching methods. The method comments in 'get_struct_category' were reformatted for better clarity and understanding.