Refactor get_elastic_data to use specific indices
Removed querying of generic index SV_INDEX via search method in get_elastic_data function in view_cbv.py. Instead, we are now directly getting a document from an index specific to the type of document (SV_LAB_INDEX in case of "lab") using the get method. This enhances query performance and accuracy as we are focusing on more specific and relevant data.