Refactor entity retrieval and validation in view_cbv.py
Changes made to enhance accuracy and improve navigation in the system by modifying the referencing and
retrieval of entities from the database. Instead of 'halStructId', 'docid' is now used for confirmation
parameters and credential forms. In addition, 'sovisu_validated' replaces 'validated' in the validation
process. Error handling added to the entity data retrieval for better UX.