Refactored context data and validated_notices_state methods
The code has been simplified by reducing redundancy and increasing code abstraction. The get_context_data method is moved to the end of the CommonContextMixin class for better organization. The ElasticContextMixin class was removed and its method 'validated_notices_state' was moved into the CommonContextMixin class with a parameter name change from 'entity' to 'p_id'. Several classes that inherited the ElasticContextMixin now only inherit the CommonContextMixin class. This change was made to improve code maintainability and readability.