Refactor Elasticsearch query in view_cbv.py
The code for Elasticsearch's count and search operations were refactored. The operations were repeated for different index types. We have now made it DRY by keeping the operations universal. The difference between data sources is handed by assigning the right index to a variable, depending on the data type. This change improves maintainability and makes the code cleaner and easier to understand.