Replace vizualisation_url function with constant
The function vizualisation_url was replaced with a constant, KIBANA_URL. This function was only returning a hard-coded URL and was unnecessary. Using a constant makes the code cleaner and gives better performance. The constant KIBANA_URL was added in constants.py and all occurrences of vizualisation_url function calls were replaced with this constant in views_cbv.py. The function vizualisation_url was then removed from viewsActions.py.