Refactor get_state_case method and update check.html
Improved the get_state_case method in view_cbv.py to more accurately retrieve researchers by strictly checking for the "sovisu_category" value. Also optimized the query to check for matches in the researcher's affiliation, thus reducing the likelihood of false positives.
In check.html, adjusted the condition within an if statement to more accurately reflect the type of entity being managed. This change may prevent unnecessary or incorrect execution of the subsequent operations.