Simplify entity name display in searchers_index.html
Reduced verbosity in displaying entity names on searchers_index.html. Instead of checking for both entity.ldapId and entity.halStructId, we now just display entity's first name initial and last name. This simplifies the code, and is based on the assumption that all entities have these attributes. This change should make the user interface cleaner and easier to read.