Update sovisuhal's ElasticSearch filter and Kibana Dashboard
Switched from filterRsr and filterLab to filtre_idhal. The change makes the filtering based on idhal resulting in more consistency as idhal is unique for each entity. Also updated Kibana's Dashboard to reflect these changes.

Removed unused function get_elastic_data(). This function previously generated the Rsr and Lab filters. Now that we're using idhal, this function is no longer needed.

Added new terminologie_lexicon_lab.ndjson file as part of an update to the Kibana dashboard. It will provide the views of the lexicon related to a lab.