Add dynamic Kibana URL to the context of View
The hard-coded Kibana URL in the iframe of the search2.html was replaced with a dynamic URL that's passed as context to the View. This change allows us to configure the Kibana URL in one place, which is particularly useful when we move from a development environment to a production environment.